requirements.txt
boto3==1.16.4
pydantic==1.6.1
chalice==1.21.2
newspaper3k==0.2.6